# Conflicts:
#	dtm-cache/pom.xml
#	dtm-calcite-adqm/pom.xml
#	dtm-calcite-core/pom.xml
#	dtm-common/pom.xml
#	dtm-jdbc-driver/pom.xml
#	dtm-kafka-core/pom.xml
#	dtm-query-execution-core/pom.xml
#	dtm-query-execution-plugin-api/pom.xml
#	dtm-status-monitor/pom.xml
#	plugins/dtm-query-execution-plugin-adb/pom.xml
#	plugins/dtm-query-execution-plugin-adg/pom.xml
#	plugins/dtm-query-execution-plugin-adp/pom.xml
#	plugins/dtm-query-execution-plugin-adqm/pom.xml
#	plugins/pom.xml
#	pom.xml

+++ dtm-jdbc-driver/src/test/java/io/arenadata/dtm/jdbc/ext/DtmResultSetMetaDataTest.java
@@ -24,8 +24,6 @@ import org.junit.jupiter.api.Test;
 import java.sql.JDBCType;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
@@ -71,7 +69,7 @@ class DtmResultSetMetaDataTest {
     @Test
     void getColumnType() throws SQLException {
         int[] expectedTypes = new int[]{
-                JDBCType.INTEGER.getVendorTypeNumber(),
+                JDBCType.BIGINT.getVendorTypeNumber(),
                 JDBCType.INTEGER.getVendorTypeNumber(),
                 JDBCType.BIGINT.getVendorTypeNumber(),
                 JDBCType.BOOLEAN.getVendorTypeNumber(),
